Improved loop support and refactoring.
IAudioSource now has SetLoopStart and SetLoopEnd, which set a constraints on how the source should read the information based on the source's position.

Fixed a memory leak when using IAudioBuffer with XAudio2 driver.

Moved ICallback to IAudioBufferCallback and moved it from CreateBuffer parameters to IAudioBuffer::SetCallback.

Added some missing documentation.